- Queries updated to be correct - both actions implemented - unit tests for actions - added reservation data to sql sample data - cellphone attribute in customer changed to LONG

+++ SQL/setuptables.sql
@@ -1,5 +1,5 @@
 create table Customer(
-cellphone INTEGER,
+cellphone LONG,
 name varchar(30),
 address varchar(60),
 dlicense INTEGER,
